greyfish-core/src/test/java/org/asoem/greyfish/core/agent/GreyfishExpressionCallbackTest.java
/*
 * Copyright (C) 2015 The greyfish authors
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

package org.asoem.greyfish.core.agent;

import com.google.inject.Guice;
import com.google.inject.Inject;
import org.asoem.greyfish.core.eval.GreyfishExpressionFactory;
import org.asoem.greyfish.core.inject.CoreModule;
import org.asoem.greyfish.core.utils.GreyfishExpressionCallback;
import org.asoem.greyfish.utils.persistence.Persisters;
import org.junit.Test;

import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertThat;


public class GreyfishExpressionCallbackTest {

    @Inject
    private GreyfishExpressionFactory expressionFactory;

    public GreyfishExpressionCallbackTest() {
        Guice.createInjector(new CoreModule()).injectMembers(this);
    }

    @Test
    public void testPersistence() throws Exception {
        // given
        final GreyfishExpressionCallback<Object, Double> callback =
                GreyfishExpressionCallback.doubleExpression(expressionFactory.compile("1.0"));

        // when
        final GreyfishExpressionCallback<Object, Double> copy = Persisters.copyAsync(callback, Persisters.javaSerialization());

        // then
        assertThat(copy, is(equalTo(callback)));
    }
}
